Crate switch_selection

Switch selection tool

Welcome

Welcome to the documentation of the switch selection tool! The switch selection tool was developed to demonstrate the possibilities of tree-decomposition based dynamic programming algorithms for problems arising in power systems, or more specifically, distribution grids. The program was written as a part of our submission to PSCC 2024 and is primarily intended for scientific research.

Purpose of documentation

This documentation is not a manual on how to use or compile the program. For that, see the README.md file in our repository.

Instead, this documentation provides an insight into the internal structure of our code, the relationships between different structs, modules, traits, etc., and explains what each element does. If you want to know more about the algorithm of a particular function, you can look at its source code. The source code contains comments explaining what is done and why.

How it works

Here’s how this program works.

The program contains multiple key components that play different roles:

Graph — this struct is used to store networks and perform any kind of general graph-related operations with them, like adding a vertex or counting the number of connected components.

PrimitiveDGGraph — this struct is used to store specifically primitive DG-graphs and provide quick access to the data related exclusively to the SwitchSelection problem. Internally, this structure is a wrapper for a Graph object.

BaseSolver — this trait declares the essential functionality for a SwitchSelection solver. There’re two structs that implement this trait:

TreeDecompositionSolver — this struct implements the dynamic programming algorithm that we’ve proposed in our paper.

CPLEXSolver — this struct allows to solve the SwitchSelection as a MILP with CPLEX.

At the beginning, the input file is read into an instance of Graph. Next, this instance is wrapped into an instance of PrimitiveDGGraph. During the creation of the wrapper, the algorithm checks whether the input graph is topologically a primitive DG-graph and whether it contains all the attributes necessary to define an instance of the SwitchSelection problem. Aferwards, this wrapper is handed to a solver object. The solver solves the problem and then allows to access an optimal solution.

Modules

dg_graph 🔒

Data structure for primitive DG-graphs

graph 🔒

Data structure for graphs and other relevant things

solver 🔒

Different solvers for the SwitchSelection problem

Macros

pretty_panic 🔒

Pretty alternative to panic!(...)

pretty_unwrap 🔒

Pretty alternative to ...unwrap()

Functions

main 🔒