Struct switch_selection::graph::Graph

pub struct Graph {
    pub attributes: AttributeMap,
    vertices: HashMap<u64, Vertex>,
    neighbourhoods: HashMap<u64, HashSet<u64>>,
    edges: HashMap<(u64, u64), Edge>,
    next_free_vertex_id: u64,
}

Graph

Brief

This struct represents a graph made of a finite set of vertices connected with each other by edges.

Features

(Un)directed edges - the structure supports both undirected and directed edges. There can be 2 undirected edges between any 2 vertices (one each way), however, if there’s an undirected edge between 2 vertices, there can be no directed edges between them.

Attributes - vertices, edges, and a graph itself can have attributes. Attributes are named, their names are stored as Strings. Most importantly, each attribute can have a value of any type that implements traits Any, Debug, Clone, Send, and Sync at the same time! These constraints are satisfied, for example, for all primitive types.

Examples

Example 1. Basics

let mut g: Graph = Graph::new(); // create a new empty graph
g.add_vertex(Some(1));
g.add_vertex(None); // automatically chooses the ID for a new vertex (0 in this case)
g.add_vertex(None); // automatically chooses the ID for a new vertex (2 in this case)
g.add_edge(1, 0, true); // add a directed edge from 1 to 0
g.remove_vertex(2);
 
assert_eq!(g.degree(0).unwrap(), 1);

Example 2. Attributes

let mut g: Graph = Graph::new(); // create a new empty graph
g.add_vertex(Some(0));
g.add_vertex(Some(1));
g.add_edge(0, 1, false);
// Let's insert a new attribute "Hello" with string value "world!"
// for an edge connecting 0 and 1
g.get_edge_mut(0, 1).unwrap().attributes.add("Hello".to_string(), "world!".to_string());
// We can also use &str to specify attribute names
g.get_vertex_mut(1).unwrap().attributes.add("Data", vec![3.1415926535_f32, -2.7e5]);
 
// Since Rust is a statically typed language, we must remember the type of
// each attribute we create to be able to access its value
assert_eq!(g.get_vertex(1).unwrap().attributes.get::<Vec::<f32>, _>("Data").unwrap(), [3.1415927, -270000.0]);

Example 3. Input/output

// You can load/save graphs from/to different files.
let mut g: Graph = Graph::new(); // create a new empty graph
{
    let input: File = File::open("example.gnbs");
    g.read_gnbs(&input).unwrap();
}
g.contract_vertices(0, 1, true);
{
    let mut output: File = File::options().write(true).truncate(true).create(true).open("example.gnbs");
    g.write_gnbs(&mut output).unwrap();
}

Fields
attributes: AttributeMapvertices: HashMap<u64, Vertex>neighbourhoods: HashMap<u64, HashSet<u64>>edges: HashMap<(u64, u64), Edge>next_free_vertex_id: u64
Implementations

impl Graph

pub fn new() -> Graph

Create a new graph

Brief

Generates a new and empty graph.

Arguments

None

Returns

Graph - a new empty graph.