Trait switch_selection::graph::ImmutableGraph

pub trait ImmutableGraph {

    // Required methods
    fn check_edge(&self, id1: u64, id2: u64) -> Option<i8>;
    fn connected_components_count(&self) -> usize;
    fn degree(&self, id: u64) -> Option<usize>;
    fn degree_in(&self, id: u64) -> Option<usize>;
    fn degree_out(&self, id: u64) -> Option<usize>;
    fn edge_count(&self) -> usize;
    fn get_edge(&self, id1: u64, id2: u64) -> Option<&Edge>;
    fn get_edge_cond<T: Fn(u64, u64) -> bool>(
        &self,
        filter: T
    ) -> Option<(u64, u64, &Edge)>;
    fn get_vertex(&self, id: u64) -> Option<&Vertex>;
    fn get_vertex_children(&self, id: u64) -> Option<HashSet<u64>>;
    fn get_vertex_cond<T: Fn(u64) -> bool>(
        &self,
        filter: T
    ) -> Option<(u64, &Vertex)>;
    fn get_vertex_neighbourhood(&self, id: u64) -> Option<HashSet<u64>>;
    fn iter_edge(&self) -> Iter<'_, (u64, u64), Edge>;
    fn iter_vertex(&self) -> Iter<'_, u64, Vertex>;
    fn vertex_count(&self) -> usize;

}

Functions that cannot mutate a graph

Brief

This trait contains functions for graphs and graph containers that don’t affect their contents. All functions in this trait borrow an immutable reference to a graph they are applied to, so you can’t borrow any mutable references to it at the same time. However, you can borrow as many more immutable references to it as you need.

Panic policy

Functions in this trait never panic.

Required Methods

fn check_edge(&self, id1: u64, id2: u64) -> Option<i8>

Check the existence of an edge

Brief

Checks if there exists any edge between two vertices.

Arguments

id1 : VertexId - the ID of the first vertex.

id2 : VertexId - the ID of the second vertex.

Returns

Option<i8> - i8 is returned to indicate how id1 and id2 are connected as stated in the table below, None is returned if at least one of the vertices doesn’t exist.

Return value	Meaning

-2	There’re directed edges going both ways

-1	There’s a directed edge from id2 to id1

0	There’s no edge between id1 and id2

1	There’s a directed edge from id1 to id2

2	There’s an undirected edge between id1 and id2

fn connected_components_count(&self) -> usize

Get the number of connected components

Brief

Counts the number of connected components in the graph.

Arguments

None.

Returns

usize - the number of connected components.

Note

The function won’t work properly for graphs with directed edges.

fn degree(&self, id: u64) -> Option<usize>

Get vertex degree

Brief

Computes the number of vertices adjacent to a given vertex.

Arguments

id : VertexID - the ID of the vertex of interest.

Returns

Option<usize> - the number of vertices adjacent to the given vertex, or None if the vertex with the given ID doesn’t exist.

Note

Vertices u and v are adjacent if there’s any kind of edge between them. So, in a simple graph with two vertices u and v and a directed edge going from u to v, the degree of both u and v will be 1.

fn degree_in(&self, id: u64) -> Option<usize>

Get vertex in-degree

Brief

Computes the number of vertices adjacent to a given vertex that are connected to it by…

… an undirected edge.

… a directed edge the endpoint of which is the given vertex.

Arguments

id : VertexID - the ID of the vertex of interest.

Returns

Option<usize> - the in-degree of the given vertex, or None if the vertex with the given ID doesn’t exist.

Note

Generally speaking, graph.degree(v) == graph.degree_in(v) + graph.degree_out(v).

fn degree_out(&self, id: u64) -> Option<usize>

Get vertex out-degree

Brief

Computes the number of vertices adjacent to a given vertex that are connected to it by…

… an undirected edge.

… a directed edge the startpoint of which is the given vertex.

Arguments

id : VertexID - the ID of the vertex of interest.

Returns

Option<usize> - the out-degree of the given vertex, or None if the vertex with the given ID doesn’t exist.

Note

Generally speaking, graph.degree(v) == graph.degree_in(v) + graph.degree_out(v).

fn edge_count(&self) -> usize

Get the number of edges

Brief

Reveals the number of edges in the graph.

Arguments

None.

Returns

usize - the number of edges in the graph.

fn get_edge(&self, id1: u64, id2: u64) -> Option<&Edge>

Access an edge

Brief

Finds an edge by the IDs of the vertices incident on it and returns an immutable reference to it.

Arguments

id1 : VertexId - the ID of the first incident vertex.

id2 : VertexId - the ID of the second incident vertex.

Returns

Option<&Edge> - an immutable reference to the found edge, or None.

Note

If there’s no undirected edge between id1 and id2, Some(&Edge) will only be returned if there’s a directed edge from id1 to id2.

fn get_edge_cond<T: Fn(u64, u64) -> bool>( &self, filter: T ) -> Option<(u64, u64, &Edge)>

Access an edge satisfying a condition

Brief

Finds an edge that satisfies a given condition and returns an immutable reference to it. The condition must be checked by a callable object that accepts two VertexId’s and returns bool.

Arguments

filter : T: Fn(VertexId, VertexId) -> bool - the condition that has to be satisfied.

Returns

Option<(VertexId, VertexId, &Edge)> - a tuple containing the IDs of two adjacent vertices and an immutable reference to the found edge connecting them, or None if all edges of the graph don’t satisfy the condition.

Note

Don’t use this function to sample random edges! The return values are not random and depend entirely on the internal structure of the hash set used to store the edges.

Examples

Example 1. Basic usage

let mut g = Graph::new();
g.add_vertex(None); // adds vertex with ID 0
g.add_vertex(None); // adds vertex with ID 1
g.add_vertex(None); // adds vertex with ID 2
g.add_vertex(None); // adds vertex with ID 3
g.add_edge(0, 1, false); // adds edge 0 --- 1
g.add_edge(3, 2, true); // adds edge 2 <-- 3
 
let x = g.get_edge_cond(|id1, id2| g.degree(id1) == 1 && g.degree(id2) == 1).unwrap();
assert!((x.0, x.1) == (0, 1) || (x.0, x.1) == (3, 2));

Example 2. Incorrect sampling of multiple edges

// Let's say, we continue the previous example and now want to sample 2
// edges that connect vertices of degree 1.
// A method one could initially think of is:
let mut found_edges: HashSet<(VertexId, VertexId)> = HashSet::new();
for i in 0..=1 {
    let x = g.get_edge_cond(|id1, id2| g.degree(id1) == 1 && g.degree(id2) == 1).unwrap();
    found_edges.insert((x.0, x.1));
}
 
assert!(found_edges.contains((0, 1)) && found_edges.contains((3, 2))); // ASSERTION FAILED

Example 3. Correct sampling of multiple edges

// Instead of the code above we should write:
let mut found_edges: HashSet<(VertexId, VertexId)> = HashSet::new();
for i in 0..=1 {
    let x = g.get_vertex_cond(|id1, id2| g.degree(id1) == 1 && g.degree(id2) == 1 && !found_edges.contains(id)).unwrap();
    found_edges.insert((x.0, x.1));
}
 
assert!(found_edges.contains((0, 1)) && found_edges.contains((3, 2)));

fn get_vertex(&self, id: u64) -> Option<&Vertex>

Access a vertex

Brief

Finds a vertex by its ID and returns an immutable reference to it.

Arguments

id : VertexId - the ID of the vertex of interest.

Returns

Option<&Vertex> - an immutable reference to the found vertex, or None.

fn get_vertex_children(&self, id: u64) -> Option<HashSet<u64>>

Get vertex children

Brief

Constructs a set of IDs of all vertices adjacent to a given vertex that are connected to it by…

… an undirected edge.

… a directed edge the startpoint of which is the given vertex.

Arguments

id : VertexId - the ID of the vertex of interest.

Returns

Option<HashSet<VertexId>> - a set containing IDs of all children of the given vertex, or None if the given vertex doesn’t exist.

fn get_vertex_cond<T: Fn(u64) -> bool>( &self, filter: T ) -> Option<(u64, &Vertex)>

Access a vertex satisfying a condition

Brief

Finds a vertex that satisfies a given condition and returns an immutable reference to it. The condition must be checked by a callable object that accepts a VertexId and returns bool.

Arguments

filter : T: Fn(VertexId) -> bool - the condition that has to be satisfied.

Returns

Option<(VertexId, &Vertex)> - a tuple of the vertex ID and an immutable reference to it, or None if all vertices of the graph don’t satisfy the condition.

Note

Don’t use this function to sample random vertices! The return values are not random and depend entirely on the internal structure of the hash set used to store the vertices.

Examples

Example 1. Basic usage

let mut g = Graph::new();
g.add_vertex(None); // adds vertex with ID 0
g.add_vertex(None); // adds vertex with ID 1
g.add_vertex(None); // adds vertex with ID 2
g.add_vertex(None); // adds vertex with ID 3
g.add_edge(0, 1, false); // adds edge 0 --- 1
g.add_edge(3, 2, true); // adds edge 2 <-- 3
 
let x = g.get_vertex_cond(|id| id % 2 == 0 && g.degree(id) == 1).unwrap();
assert!(x.0 == 0 || x.0 == 2);

Example 2. Incorrect sampling of multiple vertices

// Let's say, we continue the previous example and now want to sample 2
// vertices that have even IDs and degree 1.
// A method one could initially think of is:
let mut found_vertices: HashSet<VertexId> = HashSet::new();
for i in 0..=1 {
    found_vertices.insert(g.get_vertex_cond(|id| id % 2 == 0 && g.degree(id) == 1).unwrap().0);
}
 
assert!(found_vertices.contains(0) && found_vertices.contains(2)); // ASSERTION FAILED

Example 3. Correct sampling of multiple vertices

// Instead of the code above we should write:
let mut found_vertices: HashSet<VertexId> = HashSet::new();
for i in 0..=1 {
    found_vertices.insert(g.get_vertex_cond(|id| id % 2 == 0 && g.degree(id) == 1 && !found_vertices.contains(id)).unwrap().0);
}
 
assert!(found_vertices.contains(0) && found_vertices.contains(2));

fn get_vertex_neighbourhood(&self, id: u64) -> Option<HashSet<u64>>

Get vertex neighbourhood

Brief

Constructs a set of IDs of all vertices that are adjacent to a given vertex.

Arguments

id : VertexId - the ID of the vertex of interest.

Returns

Option<HashSet<VertexId>> - a set containing IDs of all vertices adjacent to the given vertex, or None if the given vertex doesn’t exist.

Note

Vertices u and v are adjacent if there’s any kind of edge between them.

fn iter_edge(&self) -> Iter<'_, (u64, u64), Edge>

Iterate over edges

Brief

Iterates over all edges of the graph.

Arguments

None.

Returns

hash_map::Iter<'_, (VertexId, VertexId), Edge> - an iterator over all edges.

fn iter_vertex(&self) -> Iter<'_, u64, Vertex>

Iterate over vertices

Brief

Iterates over all vertices of the graph.

Arguments

None.

Returns

hash_map::Iter<'_, VertexId, Vertex> - an iterator over all vertices.

fn vertex_count(&self) -> usize

Get the number of vertices

Brief

Reveals the number of vertices in the graph.

Arguments

None.

Returns

usize - the number of vertices in the graph.