Trait switch_selection::graph::ImmutableGraphWrapper

pub trait ImmutableGraphWrapper {
    // Required method
    fn unwrap(&self) -> &Graph;
}

A wrapper for special (user-defined) classes of immutable graphs

Brief

This trait allows users to create custom structs that store at least one instance of Graph in themselves and which should themselves be treated as graphs. Thanks to the blanket implementation of ImmutableGraph for all structs that implement ImmutableGraphWrapper, all functions from ImmutableGraph will be automatically available for your wrapper.

Panic policy

Depends on the end user.

Example

// We create a special struct for binary trees
pub struct BinaryTree {
    g: Graph,
}
 
// Using this we can apply all the functions from ImmutableGraph directly to
// instances of BinaryTree
impl ImmutableGraphWrapper for BinaryTree {
    fn unwrap(&self) -> &Graph {
        self.g
    }
}
 
// Using this we can define new functions that will only be available for
// binary trees but not other graphs
impl BinaryTree {
    fn invert(&mut self) {
        // code to invert a binary tree
    }
}

Required Methods

fn unwrap(&self) -> &Graph

Unwrap the wrapper

Brief

Access the instance of the Graph structure stored inside this wrapper.

Arguments

None.

Returns

&Graph - an immutable reference to the graph stored in the wrapper.