Trait switch_selection::graph::MutableGraph

pub trait MutableGraph {
    // Required methods
    fn add_edge(&mut self, id1: u64, id2: u64, directed: bool);
    fn add_vertex(&mut self, id: Option<u64>) -> u64;
    fn contract_vertices(
        &mut self,
        primary_vertex_id: u64,
        secondary_vertex_id: u64,
        prevent_loop: bool
    );
    fn get_edge_mut(&mut self, id1: u64, id2: u64) -> Option<&mut Edge>;
    fn get_vertex_mut(&mut self, id: u64) -> Option<&mut Vertex>;
    fn iter_edge_mut(&mut self) -> IterMut<'_, (u64, u64), Edge>;
    fn iter_vertex_mut(&mut self) -> IterMut<'_, u64, Vertex>;
    fn remove_edge(&mut self, id1: u64, id2: u64) -> Option<Edge>;
    fn remove_vertex(&mut self, id: u64) -> Option<Vertex>;
}

Functions that can mutate a graph

Brief

This trait contains functions for graphs and graph containers that do affect their contents. All functions in this trait borrow a mutable reference to a graph they are applied to, so you can’t borrow any other references to it at the same time.

Panic policy

Functions in this trait never panic.

Required Methods

fn add_edge(&mut self, id1: u64, id2: u64, directed: bool)

Add a new edge

Brief

Adds a new edge to the given graph.

Arguments

id1 : VertexId - the ID of the first vertex. If directed is true, an arc will be added that starts at this vertex.

id2 : VertexId - the ID of the second vertex. If directed is true, an arc will be added that ends at this vertex.

directed : bool - identifies whether the new edge must be directed or not.

Returns

None.

Note

This function does nothing if any of the vertices doesn’t exist or if an edge between the given vertices already exists. Exception: if there’s a directed edge from id2 to id1, then a new directed edge from id1 to id2 can be added.

fn add_vertex(&mut self, id: Option<u64>) -> u64

Add a new vertex

Brief

Adds a new vertex to the given graph.

Arguments

id : Option<VertexId> - an ID of the new vertex. If None is given, the lowest possible value will be assigned to the new vertex.

Returns

VertexId - the ID of the added vertex.

Note

If a vertex with the given ID alredy exists, it will be silently replaced with a new one. Thus, all attributes of the existing vertex together with their values will be lost. However, edges incident to the existing vertex will be kept in their original form and they will become incident to the new vertex.

fn contract_vertices( &mut self, primary_vertex_id: u64, secondary_vertex_id: u64, prevent_loop: bool )

Contract vertices

Brief

Merges a secondary vertex into a primary vertex. All edges incident on the secondary vertex will become incident on the primary vertex. All attributes of the primary vertex will be saved, all attributes of the secondary vertex will be lost.

Arguments

primary_vertex_id : VertexId - the ID of the primary vertex.

secondary_vertex_id : VertexId - the ID of the secondary vertex.

prevent_loop : bool - if there’s an edge between the primary and the secondary vertex, then this edge will either be removed (when set to true) or become a loop (when set to false).

Returns

None.

Note

This function does nothing if any of the vertices doesn’t exist.

fn get_edge_mut(&mut self, id1: u64, id2: u64) -> Option<&mut Edge>

Access an edge

Brief

Finds an edge by the IDs of the vertices incident to it and returns a mutable reference to it.

Arguments

id1 : VertexId - the ID of the first incident vertex.

id2 : VertexId - the ID of the second incident vertex.

Returns

Option<&mut Edge> - a mutable reference to the found edge, or None.

Note

If there’s no undirected edge between id1 and id2, Some(&mut Edge) will only be returned if there’s a directed edge from id1 to id2.

fn get_vertex_mut(&mut self, id: u64) -> Option<&mut Vertex>

Access a vertex

Brief

Finds a vertex by its ID and returns a mutable reference to it.

Arguments

id : VertexId - the ID of the vertex of interest.

Returns

Option<&mut Vertex> - a mutable reference to the found vertex, or None.

fn iter_edge_mut(&mut self) -> IterMut<'_, (u64, u64), Edge>

Iterate over edges

Brief

Iterates over all edges of the graph.

Arguments

None.

Returns

hash_map::IterMut<'_, (VertexId, VertexId), Edge> - a mutable iterator over all edges.

fn iter_vertex_mut(&mut self) -> IterMut<'_, u64, Vertex>

Iterate over vertices

Brief

Iterates over all vertices of the graph.

Arguments

None.

Returns

hash_map::IterMut<'_, VertexId, Vertex> - a mutable iterator over all vertices.

fn remove_edge(&mut self, id1: u64, id2: u64) -> Option<Edge>

Remove edge

Brief

Removes an edge from the given graph.

Arguments

id1 : VertexId - the ID of the first vertex.

id2 : VertexId - the ID of the second vertex.

Returns

Option<Edge> - the removed edge is returned if it had existed, None is returned otherwise.

Note

If edges connecting id1 and id2 are directed, this function will only attempt to remove the edge from id1 to id2.

fn remove_vertex(&mut self, id: u64) -> Option<Vertex>

Remove vertex

Brief

Removes a vertex from the given graph together with all edges incident on it.

Arguments

id : VertexId - the ID of the vertex of interest.

Returns

Option<Vertex> - the removed vertex is returned if it had existed, None is returned otherwise.