Trait switch_selection::solver::base_solver::BaseSolver

pub trait BaseSolver: Sized {
    // Required methods
    fn with_input(input: PrimitiveDGGraph) -> Result<Self, SolverError>;
    fn get_solution(&self) -> Option<(Graph, i8)>;
    fn solve(&mut self) -> Result<(), SolverError>;
}

Basic functionality of any SwitchSelection solver

Brief

This train defines functions that must be implemented by any solver that solves the SwitchSelection problem.

Required Methods

fn with_input(input: PrimitiveDGGraph) -> Result<Self, SolverError>

Create a new solver

Brief

Creates a new solver for the given SwitchSelection instance. The instance is expected to be defined by a primitive DG-graph and its vertex and edge attributes.

Arguments

input : PrimitiveDGGraph - the SwitchSelection instance.

Returns

Result<Self, SolverError> - Ok(...) is returned when the solver was successfully created; Err(...) is returned when any kind of error occurs.

fn get_solution(&self) -> Option<(Graph, i8)>

Retrieve an optimal solution

Brief

Constructs an optimal solution if the problem has already been solved.

Arguments

None.

Returns

Option<(Graph, TapValue)> - Some(...) is returned if the problem has been solved and there exists a solution; the Graph instance contains an optimal set of tap values and switch positions, while TapValue shows the optimal value of the objective function; None is returned if the problem has been found infeasible or if the problem hasn’t been solved yet.

fn solve(&mut self) -> Result<(), SolverError>

Solve the problem

Brief

Solves the SwitchSelection problem instance.

Arguments

None.

Returns

Result<(), SolverError> - Ok(...) is returned if the problem was successfully solved; Err(...) is returned in the case of any error or if the problem was found infeasible.