Function switch_selection::solver::benchmark::start_benchmark

pub fn start_benchmark(
    treewidth: u8,
    min_junction_count: u8,
    max_junction_count: u8,
    sample_count: u8,
    sample_repeat: u8,
    sample_ignore: u8
) -> Result<(), SolverError>

Initiate the benchmark

Brief

Starts the benchmark of the TreeDecompositionSolver and the CPLEXSolver on a series of randomly generated SwitchSelection instances.

Arguments

treewidth : u8 - the treewidth of primitive DG-graphs, on which the solvers have to be tested.

min_junction_count : u8 - the minimum number of junctions in the random primitive DG-graphs.

max_junction_count : u8 - the maximum number of junctions in the random primitive DG-graphs.

sample_count : u8 - the number of primitive DG-graphs/SwitchSelection instances to generate for each number of junctions.

sample_repeat : u8 - the number of times each instance has to be solved by each solver.

sample_ignore : u8 - indicates how many first repeats of solving each instance should be ignored when computing metrics.

Returns

Result<(), SolverError> - Ok(...) is returned when the benchmark is successfully completed; Err(...) is returned when any kind of error occurs.

Note

During its execution, this function will print messages to the console that show, in particular, the time each solver spends to solve each instance of SwitchSelection. This time is computed as the average running time of the solver over sample_repeat - sample_ignore last repeats.