Struct switch_selection::solver::tree_decomposition_solver::DirectedTreeDecomposition

struct DirectedTreeDecomposition {
    core: Graph,
}

Directed tree decomposition

Brief

This struct represents a directed tree decomposition used by our dynamic programming algorithm. This struct is a wrapper around Graph.

Fields
core: Graph
Implementations

impl DirectedTreeDecomposition

pub fn new(input: &Graph) -> Result<DirectedTreeDecomposition, SolverError>

Create a directed tree decomposition

Brief

Constructs a new directed tree decomposition for the given graph.

Arguments

input : &Graph - the graph to construct the tree decomposition for.

Returns

Result<DirectedTreeDecomposition, SolverError> - Ok(...) is returned if the directed tree decomposition has been successfully constructed; Err(...) is returned in case of any error.

pub fn dfs_postordering(&self) -> Vec<u64>

Get a DFS post-ordering

Brief

Constructs a depth-first search vertex post-ordering to use in the dynamic programming algorithm as the order of bag traversal.

Arguments

None.

Returns

Vec<VertexId> - the post-ordering.

pub fn get_root_id(&self) -> u64

Get the root ID

Brief

Gets the ID of the root bag.

Arguments

None.

Returns

VertexId - the ID of the root bag.