Struct switch_selection::solver::tree_decomposition_solver::TapsMemo

struct TapsMemo {
    primary_substations: Vec<u64>,
    table: HashSet<TapsMemoEntry>,
}

Memo

Brief

This struct defines a memo that is stored for each bag of the tree decomposition. They’re used for backtracking an optimal solution after the solver has finished its work.

Fields
primary_substations: Vec<u64>table: HashSet<TapsMemoEntry>
Implementations

impl TapsMemo

pub fn new(primary_substations: Vec<u64>) -> TapsMemo

Create a new memo

Brief

Creates a new empty memo for the given list of primary substations.

Arguments

primary_substations : Vec<VertexId> - primary substations.

Returns

TapsMemo - a new memo.

pub fn intersect(&mut self, other: &TapsMemo)

Intersect two memos

Brief

Keeps only those entries in the memo that have at least one corresponding entry in another memo. Also modifies the stored values of the objective function accordingly.

Arguments

other : &TapsMemo - a memo of another bag of the tree decomposition.

Returns

None.