Struct switch_selection::solver::tree_decomposition_solver::TreeDecompositionSolver

pub struct TreeDecompositionSolver {
    input: Arc<PrimitiveDGGraph>,
    dtd: Arc<DirectedTreeDecomposition>,
    memos: Option<HashMap<u64, TapsMemo>>,
    thread_count: usize,
}

TreeDecompositionSolver

Brief

TreeDecompositionSolver solves the SwitchSelection using dynamic programming with memoisation. We deploy a directed tree decomposition of the underlying primitive DG-graph as the execution graph of our algorithm.

Fields
input: Arc<PrimitiveDGGraph>dtd: Arc<DirectedTreeDecomposition>memos: Option<HashMap<u64, TapsMemo>>thread_count: usize
Implementations

impl TreeDecompositionSolver

fn thread_workload( input: Arc<PrimitiveDGGraph>, memos: Arc<Mutex<HashMap<u64, TapsMemo>>>, dtd: Arc<DirectedTreeDecomposition>, bag_id: u64, rx: Receiver<u64> ) -> Result<(), SolverError>

Workload for a single compute thread

Brief

Executed by every compute thread spawned by the main thread. Does brute force enumeration of all possible tap positions for one bag.

Arguments

input : Arc<PrimitiveDGGraph> - a SwitchSelection instance.

memos : Arc<Mutex<HashMap<VertexId, TapsMemo>>> - a collection of memos for all bags of the tree decomposition.

dtd : Arc<DirectedTreeDecomposition> - a directed tree decomposition of the input graph.

bag_id : VertexId - the ID of the bag that this thread has to process.

rx : Receiver<VertexId> - an MPSC receiver that allows this thread to receive messages from the main thread.

Returns

Result<(), SolverError> - Ok(...) is returned when the computations are successfully completed; Err(...) is returned in case of any error.